fix(geometry): allocate axis_angle_to_quaternion output from the computed values
The output buffer took its dtype from the input, so an integer axis angle got
the float results written back into an integer tensor and every component
truncated to zero: axis_angle_to_quaternion(torch.tensor([1, 0, 0])) returned
tensor([0, 0, 0, 0]) with no warning.

torch.sqrt already promotes an integer input to float32, so allocating the
buffer with the dtype of the computed values returns the promoted result
instead of silent zeros. Float inputs are untouched, since the computed dtype
is the input dtype for float16, bfloat16, float32 and float64.

+++ kornia/geometry/conversions.py
@@ -918,9 +918,10 @@ def axis_angle_to_quaternion(axis_angle: torch.Tensor) -> torch.Tensor:
     k: torch.Tensor = torch.where(mask, k_pos, k_neg)
     w: torch.Tensor = torch.where(mask, torch.cos(half_theta), ones)
 
-    quaternion: torch.Tensor = torch.zeros(
-        size=(*axis_angle.shape[:-1], 4), dtype=axis_angle.dtype, device=axis_angle.device
-    )
+    # the output dtype comes from the computed values, not from the input: an integer input is
+    # promoted to floating point by the sqrt above, and writing that back into an integer buffer
+    # would truncate every component to zero.
+    quaternion: torch.Tensor = torch.zeros(size=(*axis_angle.shape[:-1], 4), dtype=w.dtype, device=axis_angle.device)
     quaternion[..., 1:2] = a0 * k
     quaternion[..., 2:3] = a1 * k
     quaternion[..., 3:4] = a2 * k
